requirements.txt
faiss-cpu
numpy
sentence-transformers
tqdm
openai
pyyaml
